API/requirements.txt
flask
PasteScript==3.2.0
waitress
docker
werkzeug
